requirements.txt
alembic==1.0.7
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
WTForms==2.2.1
gunicorn
Flask-LoginBcrypt==3.1.6flask-bcrypt==0.7.1
